Move budget code into useBudget file
