+++ web/src/pages/budget/budget.vue
@@ -7,7 +7,7 @@
   <UBadge :label="(new Date(budget.end_date.toString())).toDateString()" />
 </h3>
 <h3>TOTAL BALANCE:
-  <DollarAmount :amount="calculateTotalBalance" />
+  <DollarAmount :amount="budget_balance" />
   <UBadge label="Out of" />
   <DollarAmount :amount="budget.amount" />
 </h3>
@@ -22,22 +22,14 @@
 </template>
 
 <script lang="ts" setup>
-import useBudget from '~/composables/useBudget';
+import useBudget, { get_budget_transactions, calculateTotalBalance } from '~/composables/useBudget';
 import useTransactions, { TransactionType, type Transaction } from '~/composables/useTransactions';
+
+
 const transactions = useTransactions()
 const budget = useBudget()
-
-const get_transactions = ((): Transaction[] => {
-  let new_transactions = []
-  for (const transaction of transactions.value) {
-    if (transaction.date >= budget.value.start_date && transaction.date <= budget.value.end_date) {
-      new_transactions.push(transaction)
-    }
-
-  }
-  return new_transactions
-})
-const transactions_in_budget: Transaction[] = get_transactions()
+const transactions_in_budget: Transaction[] = get_budget_transactions(transactions.value, budget.value)
+const budget_balance = calculateTotalBalance(transactions_in_budget, Number(budget.value.amount))
 
 const columns = [{
   key: 'store',
@@ -54,17 +46,5 @@ const columns = [{
   sortable: true,
 }]
 
-const calculateTotalBalance = computed(() => {
-  let total = 0
-
-  for (const transaction of transactions_in_budget) {
-    if (transaction.type === TransactionType.EXPENSE)
-      total -= Number(transaction.amount)
-    else if (transaction.type === TransactionType.INCOME)
-      total += Number(transaction.amount)
-  }
-
-  return Number(budget.value.amount) + total
-})
 
 </script>
